# backend/maintenance-service/openapi.yaml
openapi: 3.0.3
info:
  title: Maintenance Service API
  description: |
    Maintenance microservice for Orchid Vehicle Rental System.
    
    ## Features
    - Issue reporting and tracking
    - Maintenance schedule management
    - Vehicle maintenance history
    - Email notifications for issue reports
    - Role-based access control
    
    ## Authentication
    Postman usage is through API Gateway with Bearer JWT in `Authorization`.
    The gateway/JWT filter can propagate these optional headers to this service:
    - `X-User-Role`: User role (CUSTOMER, REPAIR_ADVISOR, OWNER)
    - `X-User-Id`: User identifier
    - `X-Username`: Username of the user
    
    ## Inter-service Communication
    - **Vehicle Service**: Validates vehicle existence and updates status
    - **Auth Service**: Retrieves user information and email
    - **Notification Service**: Sends email notifications for reported issues
    - **Payment Service**: Processes damage payments
    
  version: 1.0.0
  contact:
    name: Orchid Vehicle Rental Team
    email: support@orchid-rental.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080
    description: Local API Gateway URL (recommended for Postman testing)
  - url: http://localhost:8085
    description: Local Development Server
  - url: http://api-gateway:8080
    description: API Gateway (Via Gateway)
  - url: https://maintenance-service.azurecontainerapps.io
    description: Production Server (Azure Container Apps)

security:
  - bearerAuth: []

tags:
  - name: Issue Reporting
    description: Report and manage vehicle issues
  - name: Maintenance Scheduling
    description: Schedule and manage maintenance operations
  - name: Maintenance Records
    description: View maintenance history and records
  - name: Maintenance Operations
    description: Start, complete, and update maintenance status
  - name: Administration
    description: Utility and delete operations

paths:
  /api/maintenance/issues/report:
    post:
      tags:
        - Issue Reporting
      summary: Report a vehicle issue
      description: |
        Creates a new maintenance issue report for a vehicle.
        Only CUSTOMER and REPAIR_ADVISOR roles can report issues.
        Email notification is automatically sent to the user.
      parameters:
        - $ref: '#/components/parameters/XUserRole'
        - $ref: '#/components/parameters/XUserId'
        - $ref: '#/components/parameters/XUsername'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportedIssue'
      responses:
        '200':
          description: Issue reported successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportedIssue'
        '400':
          description: Invalid input or vehicle not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied - insufficient permissions

  /api/maintenance/issues/vehicle/{vehicleId}:
    get:
      tags:
        - Issue Reporting
      summary: Get issues for a vehicle
      description: Retrieves all maintenance issues reported for a specific vehicle.
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
          description: Vehicle ID (MongoDB ObjectId)
      responses:
        '200':
          description: List of issues for the vehicle
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportedIssue'
        '404':
          description: Vehicle not found

  /api/maintenance/issues/{id}:
    get:
      tags:
        - Issue Reporting
      summary: Get issue details
      description: Retrieves detailed information about a specific maintenance issue.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Issue ID (MongoDB ObjectId)
      responses:
        '200':
          description: Issue details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportedIssue'
        '400':
          description: Invalid request (including issue not found)

    delete:
      tags:
        - Administration
      summary: Delete issue
      description: Deletes a reported issue by ID. Restricted to maintenance management roles.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Issue ID
        - $ref: '#/components/parameters/XUserRole'
      responses:
        '200':
          description: Issue deleted successfully
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied

  /api/maintenance/schedule-from-issue/{issueId}:
    post:
      tags:
        - Maintenance Scheduling
      summary: Schedule maintenance from issue
      description: |
        Creates a maintenance record from a reported issue.
        Only REPAIR_ADVISOR role can schedule maintenance.
      parameters:
        - name: issueId
          in: path
          required: true
          schema:
            type: string
          description: Issue ID
        - $ref: '#/components/parameters/XUserRole'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleFromIssueRequest'
      responses:
        '200':
          description: Maintenance scheduled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRecord'
        '400':
          description: Invalid request (including issue not found)
        '403':
          description: Access denied - only REPAIR_ADVISOR can schedule

  /api/maintenance/scheduled:
    post:
      tags:
        - Maintenance Scheduling
      summary: Create scheduled maintenance
      description: |
        Creates a new scheduled maintenance record directly.
        Only REPAIR_ADVISOR role can create maintenance schedules.
      parameters:
        - $ref: '#/components/parameters/XUserRole'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MaintenanceRecord'
      responses:
        '200':
          description: Maintenance scheduled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRecord'
        '400':
          description: Invalid input
        '403':
          description: Access denied

  /api/maintenance/upcoming/vehicle/{vehicleId}:
    get:
      tags:
        - Maintenance Scheduling
      summary: Get upcoming maintenance for vehicle
      description: Retrieves scheduled maintenance operations for a specific vehicle.
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
          description: Vehicle ID
      responses:
        '200':
          description: List of upcoming maintenance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MaintenanceRecord'

  /api/maintenance/upcoming/all:
    get:
      tags:
        - Maintenance Scheduling
      summary: Get all upcoming maintenance
      description: Retrieves all maintenance records in SCHEDULED state.
      responses:
        '200':
          description: List of all upcoming maintenance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MaintenanceRecord'

  /api/maintenance/{id}/start:
    put:
      tags:
        - Maintenance Operations
      summary: Start maintenance
      description: Marks a maintenance record as IN_PROGRESS.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Maintenance record ID
        - $ref: '#/components/parameters/XUserRole'
      responses:
        '200':
          description: Maintenance started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRecord'
        '400':
          description: Invalid request (including record not found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied

  /api/maintenance/{id}/complete:
    put:
      tags:
        - Maintenance Operations
      summary: Complete maintenance
      description: Marks a maintenance record as COMPLETED and stores actual cost.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XUserRole'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteMaintenanceRequest'
      responses:
        '200':
          description: Maintenance completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRecord'
        '400':
          description: Invalid request (including record not found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied

  /api/maintenance/{id}/status:
    put:
      tags:
        - Maintenance Operations
      summary: Update maintenance status
      description: Updates the status field of a maintenance record.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/XUserRole'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMaintenanceStatusRequest'
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRecord'
        '400':
          description: Invalid request (including record not found)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access denied

  /api/maintenance/history/vehicle/{vehicleId}:
    get:
      tags:
        - Maintenance Records
      summary: Get maintenance history for a vehicle
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Vehicle maintenance history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MaintenanceRecord'

  /api/maintenance/history/all:
    get:
      tags:
        - Maintenance Records
      summary: Get full maintenance history
      responses:
        '200':
          description: Full maintenance history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MaintenanceRecord'

  /api/maintenance/history/vehicle/{vehicleId}/completed:
    get:
      tags:
        - Maintenance Records
      summary: Get completed maintenance for a vehicle
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Completed maintenance records for the vehicle
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MaintenanceRecord'

  /api/maintenance/generate-recurring:
    post:
      tags:
        - Administration
      summary: Generate recurring maintenance records
      description: Triggers generation of recurring maintenance entries that are due.
      parameters:
        - $ref: '#/components/parameters/XUserRole'
      responses:
        '200':
          description: Recurring maintenance generation completed
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/maintenance/records/{id}:
    delete:
      tags:
        - Administration
      summary: Delete maintenance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Maintenance record ID
        - $ref: '#/components/parameters/XUserRole'
      responses:
        '200':
          description: Maintenance record deleted successfully
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    XUserRole:
      name: X-User-Role
      in: header
      required: false
      schema:
        type: string
        enum: [CUSTOMER, REPAIR_ADVISOR, OWNER]
      description: User role for authorization

    XUserId:
      name: X-User-Id
      in: header
      required: false
      schema:
        type: string
      description: User identifier (MongoDB ObjectId)

    XUsername:
      name: X-Username
      in: header
      required: false
      schema:
        type: string
      description: Username of the authenticated user

  schemas:
    ReportedIssue:
      type: object
      description: Vehicle issue report
      required:
        - vehicleId
        - description
        - issueType
      properties:
        id:
          type: string
          description: Issue ID (MongoDB ObjectId)
          readOnly: true
        vehicleId:
          type: string
          description: Vehicle ID (MongoDB ObjectId)
        issueType:
          type: string
          description: Type of issue (e.g., Engine Problem, Brake Issue, Suspension)
          example: "Engine Problem"
        description:
          type: string
          description: Detailed description of the issue
          example: "Engine noise on startup"
        reportedDate:
          type: string
          format: date-time
          description: Date and time when issue was reported
          readOnly: true
        reportedBy:
          type: string
          description: User ID of the person who reported the issue
          readOnly: true
        reportedByEmail:
          type: string
          format: email
          description: Email address of the person who reported the issue
          readOnly: true
        status:
          type: string
          enum: [REPORTED, UNDER_REVIEW, SCHEDULED, RESOLVED]
          description: Current status of the issue
          default: REPORTED
        maintenanceRecordId:
          type: string
          description: Associated maintenance record ID if scheduled
          nullable: true

    MaintenanceRecord:
      type: object
      description: Maintenance operation record
      required:
        - vehicleId
        - maintenanceType
      properties:
        id:
          type: string
          description: Maintenance record ID (MongoDB ObjectId)
          readOnly: true
        vehicleId:
          type: string
          description: Vehicle ID (MongoDB ObjectId)
        maintenanceType:
          type: string
          description: Type of maintenance (for example SCHEDULED, REPAIR, ACCIDENT, BREAKDOWN)
        description:
          type: string
          description: Detailed maintenance description
          nullable: true
        isRecurring:
          type: boolean
          description: Whether this is part of a recurring maintenance cycle
          default: false
        recurrencePattern:
          type: string
          description: Recurrence pattern for recurring maintenance
          enum: [MONTHLY, QUARTERLY, YEARLY]
          nullable: true
        nextDueDate:
          type: string
          format: date
          description: Next due date for recurring maintenance
          nullable: true
        issueId:
          type: string
          description: Linked reported issue ID, if created from issue flow
          nullable: true
        estimatedCost:
          type: number
          format: double
          description: Estimated maintenance cost
          nullable: true
        actualCost:
          type: number
          format: double
          description: Actual maintenance cost after completion
          nullable: true
        status:
          type: string
          enum: [SCHEDULED, IN_PROGRESS, COMPLETED, CANCELLED]
          description: Current status of maintenance record
        scheduledDate:
          type: string
          format: date-time
          description: Scheduled date and time
          nullable: true
        startDate:
          type: string
          format: date-time
          description: Date and time when maintenance started
          nullable: true
        completionDate:
          type: string
          format: date-time
          description: Date and time when maintenance was completed
          nullable: true
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          nullable: true
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          nullable: true

    CompleteMaintenanceRequest:
      type: object
      description: Request payload for completing maintenance
      properties:
        actualCost:
          type: number
          format: double
          description: Final maintenance cost
          nullable: true

    ScheduleFromIssueRequest:
      type: object
      description: Request to schedule maintenance from an issue
      required:
        - scheduledDate
      properties:
        scheduledDate:
          type: string
          format: date-time
          description: Date/time to schedule maintenance
        estimatedCost:
          type: number
          format: double
          description: Estimated cost for the scheduled work
          nullable: true

    UpdateMaintenanceStatusRequest:
      type: object
      description: Request to update maintenance status
      required:
        - status
      properties:
        status:
          type: string
          enum: [SCHEDULED, IN_PROGRESS, COMPLETED, CANCELLED]
          description: New status for the maintenance record

    ErrorResponse:
      type: object
      description: Error response
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error message
        status:
          type: integer
          description: HTTP status code
        timestamp:
          type: string
          format: date-time
          description: When the error occurred